requirements.txt
arrow==1.3.0
asgiref==3.8.1
click==8.1.8
colorama==0.4.6
diff-match-patch==20241021
Django==5.2
django-cors-headers==4.7.0
django-crontab==0.7.1
django-debug-toolbar==5.1.0
django-import-export==4.3.7
django-redis==5.4.0
django-rq==3.0.0
djangorestframework==3.16.0
python-dateutil==2.9.0.post0
redis==5.2.1
rq==2.3.1
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
times==0.7
types-python-dateutil==2.9.0.20241206
tzdata==2025.2
